+++ dio/CHANGELOG.md
@@ -7,8 +7,8 @@ See the [Migration Guide][] for the complete breaking changes list.**
 
 - Improve comments.
 - Fix error when cloning `MultipartFile` from `FormData` with regression test.
+- Deprecate `MulitpartFile` constructor in favor `MultipartFile.fromStream`.
 - Add convenience method for cloning `FormData` as a whole
-- Deprecate `stream` parameter in favor of `data` in `MultipartFile`.
 
 ## 5.3.0